+++ src/db/schemas/storage.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+from decimal import Decimal
 from sqlalchemy import ForeignKey, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from src.db import Base
@@ -21,7 +22,7 @@ class Storage(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"), nullable=False)
     account_id: Mapped[int] = mapped_column(ForeignKey("accounts.id"), nullable=False)
-    amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False, default=0)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False, default=0)
 
     user = relationship("User", back_populates="storages")
     account = relationship("Account", back_populates="storage")
