+++ src/db/schemas/transaction.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+from decimal import Decimal
 from typing import Optional
 from uuid import uuid4
 from sqlalchemy import TIMESTAMP, ForeignKey, Numeric, String, Uuid, func
@@ -48,7 +49,7 @@ class Transaction(Base):
     )
     title: Mapped[str] = mapped_column(String(64), nullable=False)
     description: Mapped[str] = mapped_column(String(256), nullable=True)
-    amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     debit_account_id: Mapped[int] = mapped_column(
         ForeignKey("accounts.id"), nullable=False
     )
